Add access flag for previously warm methods
We want to know if the method was warm instead of just having a non
zero counter. This is required if we want to not compile all of the
startup methods, but still compile warm methods.

Test: test-art-host ART_TEST_JIT=true

Bug: 62200509

+++ compiler/optimizing/intrinsics.cc
@@ -146,7 +146,7 @@ void IntrinsicsRecognizer::Run() {
           Intrinsics intrinsic = static_cast<Intrinsics>(art_method->GetIntrinsic());
           if (!CheckInvokeType(intrinsic, invoke)) {
             LOG(WARNING) << "Found an intrinsic with unexpected invoke type: "
-                << intrinsic << " for "
+                << static_cast<uint32_t>(intrinsic) << " for "
                 << art_method->PrettyMethod()
                 << invoke->DebugName();
           } else {
